waveform: Rename _timing._base -> _timing._timing
