+++ src/nitypes/waveform/_timing/__init__.py
@@ -1,7 +1,7 @@
 """Waveform timing data types for NI Python APIs."""
 
-from nitypes.waveform._timing._base import Timing
 from nitypes.waveform._timing._sample_interval import SampleIntervalMode
+from nitypes.waveform._timing._timing import Timing
 from nitypes.waveform._timing._types import (
     _AnyDateTime,
     _AnyTimeDelta,
