+++ src/nitypes/waveform/_timing/_sample_interval/_irregular.py
@@ -21,7 +21,7 @@ from nitypes.waveform._timing._types import (
 )
 
 if TYPE_CHECKING:
-    from nitypes.waveform._timing._base import Timing  # circular import
+    from nitypes.waveform._timing._timing import Timing  # circular import
 
 
 class _Direction(Enum):
