+++ src/nitypes/waveform/_timing/_timing.py
